Added a 'Who's Online' box under the Oneliner, it refreshes every 2 mins. Site owners can tweak the cache setting by copying webview/base/base.html to their local folder and adjusting.
Adjusted default artist template to use the artists flag instead of an icon, if it exists. If not, then the default icon will just be used. This stops an icon and a flag being used (saving space). Deceased flag will still show at the end of the artist name.

User template tweaked to remove littleman (noone used it anyways) and a space removal between the flag and the name.

Deleted 5 old images that were corrupt from Mercurial; They aren't used anyways.