Jinja2 implented (easy-install jinja2), now renders two pages : pm_send.html and queue_list.html Caching is removed on jinja pages (so far), and some functionality is gone.
Created a shim that check if the template exists in jinja's folder. If not it falls
back to standard django rendering. Replaced all render_to_response calls in webview/views.py
to the new one.

Also changed dv_extend a bit, removed tabs and changed the bbcode part to be more effective.

This branch should be considered highly experimantal atm :)